DatabaseManagementViewModel.cs added

+++ VRK_WPF/MVVM/ViewModel/AdminViewModels/FileRowViewModel.cs
@@ -10,7 +10,7 @@ public partial class FileRowViewModel : ObservableObject, IModifiableRow
     [ObservableProperty] private DateTime _creationTime;
     [ObservableProperty] private DateTime _modificationTime;
     [ObservableProperty] private string? _contentType;
-    [ObservableProperty] private int _chunkSize;
+    [ObservableProperty] private long _chunkSize;
     [ObservableProperty] private int _totalChunks;
     [ObservableProperty] private int _state;
     [ObservableProperty] private bool _isModified;
@@ -18,7 +18,7 @@ public partial class FileRowViewModel : ObservableObject, IModifiableRow
     partial void OnFileNameChanged(string value) => IsModified = true;
     partial void OnFileSizeChanged(long value) => IsModified = true;
     partial void OnContentTypeChanged(string? value) => IsModified = true;
-    partial void OnChunkSizeChanged(int value) => IsModified = true;
+    partial void OnChunkSizeChanged(long value) => IsModified = true;
     partial void OnTotalChunksChanged(int value) => IsModified = true;
     partial void OnStateChanged(int value) => IsModified = true;
 }
